+++ public/components/overview/compliance-table/compliance-table.tsx
@@ -37,7 +37,7 @@ export const ComplianceTable = withAgentSupportModule(class ComplianceTable exte
     _history: { history: { items: { from: string; to: string }[] } };
   };
 
-  KibanaServices: { [key: string]: any };
+  PluginPlatformServices: { [key: string]: any };
   filterManager: FilterManager;
   indexPattern: any;
   state: {
@@ -54,9 +54,9 @@ export const ComplianceTable = withAgentSupportModule(class ComplianceTable exte
 
   constructor(props) {
     super(props);
-    this.KibanaServices = getDataPlugin().query;
-    this.filterManager = this.KibanaServices.filterManager;
-    this.timefilter = this.KibanaServices.timefilter.timefilter;
+    this.PluginPlatformServices = getDataPlugin().query;
+    this.filterManager = this.PluginPlatformServices.filterManager;
+    this.timefilter = this.PluginPlatformServices.timefilter.timefilter;
     this.state = {
       selectedRequirement: '',
       flyoutOn: true,