+++ public/components/visualize/wz-visualize.js
@@ -41,7 +41,7 @@ import { compose } from 'redux';
 import { UI_LOGGER_LEVELS } from '../../../common/constants';
 import { UI_ERROR_SEVERITIES } from '../../react-services/error-orchestrator/types';
 import { getErrorOrchestrator } from '../../react-services/common-services';
-import { satisfyKibanaVersion } from '../../../common/semver';
+import { satisfyPluginPlatformVersion } from '../../../common/semver';
 
 const visHandler = new VisHandlers();
 
@@ -130,7 +130,7 @@ export const WzVisualize = compose(
         // Known fields are refreshed only once per dashboard loading
         try {
           this.setState({ hasRefreshedKnownFields: true, isRefreshing: true });
-          if(satisfyKibanaVersion('<7.11')){
+          if(satisfyPluginPlatformVersion('<7.11')){
             await PatternHandler.refreshIndexPattern(this.newFields);
           };
           this.setState({ isRefreshing: false });
@@ -157,7 +157,7 @@ export const WzVisualize = compose(
     };
     reloadToast = () => {
       const toastLifeTimeMs = 300000;
-      if(satisfyKibanaVersion('<7.11')){
+      if(satisfyPluginPlatformVersion('<7.11')){
         getToasts().add({
           color: 'success',
           title: 'The index pattern was refreshed successfully.',
@@ -172,7 +172,7 @@ export const WzVisualize = compose(
           </EuiFlexGroup>),
           toastLifeTimeMs
         });
-      }else if(satisfyKibanaVersion('>=7.11')){
+      }else if(satisfyPluginPlatformVersion('>=7.11')){
         getToasts().add({
           color: 'warning',
           title: 'Found unknown fields in the index pattern.',