+++ public/utils/config-equivalences.js
@@ -1,4 +1,4 @@
-import { ASSETS_BASE_URL_PREFIX } from "../../common/constants";
+import { ASSETS_BASE_URL_PREFIX, PLUGIN_PLATFORM_NAME } from "../../common/constants";
 
 export const configEquivalences = {
   pattern: 'Default index pattern to use on the app.',
@@ -16,11 +16,11 @@ export const configEquivalences = {
   'checks.fields':
     'Enable or disable the known fields health check when opening the app.',
     'checks.metaFields':
-      'Change the default value of the Kibana metaField configuration',
+      `Change the default value of the ${PLUGIN_PLATFORM_NAME} metaField configuration`,
     'checks.timeFilter':
-      'Change the default value of the Kibana timeFilter configuration',
+      `Change the default value of the ${PLUGIN_PLATFORM_NAME} timeFilter configuration`,
     'checks.maxBuckets':
-      'Change the default value of the Kibana max buckets configuration',
+      `Change the default value of the ${PLUGIN_PLATFORM_NAME} max buckets configuration`,
   'extensions.pci': 'Enable or disable the PCI DSS tab on Overview and Agents.',
   'extensions.gdpr': 'Enable or disable the GDPR tab on Overview and Agents.',
   'extensions.audit': 'Enable or disable the Audit tab on Overview and Agents.',