+++ public/utils/wz-logo-menu.js
@@ -10,7 +10,7 @@
  * Find more information about this on the LICENSE file.
  */
 
-// Remove Kibana Wazuh name and breadcrumb
+// Remove plugin platform Wazuh name and breadcrumb
 export const changeWazuhNavLogo = () => {
   const interval = setInterval(() => {
     const nav = document.querySelector('[data-test-subj="breadcrumbs"]  > .euiBreadcrumb');